changed style, fixed elo floats

+++ leaderboard/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Player(models.Model):
 	player_name = models.CharField(max_length=200)
-	elo = models.IntegerField(default=1500)
+	elo = models.FloatField(default=1500.0)
 
 	def __str__(self):
 		return self.player_name
